+++ .env
@@ -7,6 +7,8 @@ PORT=8080
 # list of service names
 SERVICES=("home" "news" "shop" "travel" "dsp" "dsp-a" "dsp-b" "ssp" "ssp-a" "ssp-b" "idp" "topics" "topics-server" "ad-server")
 
+DEMO_HOST_PREFIX=privacy-sandbox-demos-
+
 # home
 HOME_HOST=privacy-sandbox-demos-home.dev
 HOME_TOKEN=""
@@ -17,7 +19,6 @@ HOME_DETAIL="Home page of Privacy Sandbox Demos"
 NEWS_HOST=privacy-sandbox-demos-news.dev
 NEWS_TOKEN=""
 NEWS_DETAIL="Publisher: News media site"
-
 
 # Advertizer
 ## shop
