+++ services/ad-tech/package.json
@@ -20,6 +20,7 @@
   },
   "devDependencies": {
     "@types/cors": "^2.8.17",
+    "@types/ejs": "^3.1.5",
     "@types/express": "^4.17.17",
     "@types/node": "^20.4.2",
     "prettier": "^3.0.0",
