+++ services/ad-tech/src/lib/constants.ts
@@ -16,6 +16,7 @@ export const {
   PORT,
   EXTERNAL_PORT,
   HOSTNAME,
+  DEMO_HOST_PREFIX,
 
   // Advertisers
   SHOP_HOST,
@@ -33,44 +34,59 @@ export const {
   // Ad-techs
   DSP_HOST,
   DSP_DETAIL,
+  DSP_URI,
 
   DSP_A_HOST,
   DSP_A_DETAIL,
+  DSP_A_URI,
 
   DSP_B_HOST,
   DSP_B_DETAIL,
+  DSP_B_URI,
 
   SSP_HOST,
   SSP_DETAIL,
+  SSP_URI,
 
   SSP_A_HOST,
   SSP_A_DETAIL,
+  SSP_A_URI,
 
   SSP_B_HOST,
   SSP_B_DETAIL,
+  SSP_B_URI,
 
-  HOME_HOST,
   AD_SERVER_HOST,
+  AD_SERVER_DETAIL,
+  AD_SERVER_URI,
 } = process.env;
 
 export const CURRENT_ORIGIN = new URL(
   `https://${HOSTNAME}:${EXTERNAL_PORT}`,
 ).toString();
 
-export const DSP_ORIGIN = new URL(
-  `https://${DSP_HOST}:${EXTERNAL_PORT}`,
-).toString();
-export const DSP_A_ORIGIN = new URL(
-  `https://${DSP_A_HOST}:${EXTERNAL_PORT}`,
-).toString();
-export const DSP_B_ORIGIN = new URL(
-  `https://${DSP_B_HOST}:${EXTERNAL_PORT}`,
-).toString();
-
-/** Name of the contextual advertiser. */
+// ****************************************************************************
+// CONTEXTUAL AUCTION METADATA
+// ****************************************************************************
+/**
+ * Name of the contextual advertiser.
+ * This is used in the buyer's bid response to contextual bid requests.
+ */
 export const ADVERTISER_CONTEXTUAL = 'ContextNext';
+/** Max bid CPM for contextual auctions. */
+export const MAX_CONTEXTUAL_BID = 1.5;
+/** Min bid CPM for contextual auctions. */
+export const MIN_CONTEXTUAL_BID = 0.5;
 
-/** Shop item labels indexed by their ID. */
+// ****************************************************************************
+// ADVERTISER METADATA
+// ****************************************************************************
+/**
+ * Map of SHOP advertiser's item metadata. Map contains item labels indexed by
+ * itemId.
+ * This is currently used to initialize the sellers's BYOS key-value store for
+ * the Protected Audience seller's realtime creative scoring signals use-case.
+ */
 export const KNOWN_SHOP_ITEM_LABELS_BY_ID = {
   '1f45e': "Man's brown shoe",
   '1f45f': 'Blue running shoe',
@@ -114,3 +130,24 @@ export const MACRO_DISPLAY_RENDER_URL_AD_SIZE =
  */
 export const MACRO_VIDEO_RENDER_URL_SSP_VAST =
   'sspVast=${SSP_VAST}&sspVastAlt=%%SSP_VAST%%';
+
+// ****************************************************************************
+// INTEGRATION CONFIGURATIONS
+// ****************************************************************************
+/**
+ * Map of buyers to integrate with indexed by sellers.
+ * This is meant to showcase business relationships between buyers and sellers.
+ * When a buyer is integrated with the seller, the seller sends contextual bid
+ * requests to the buyer and includes the buyer as an interest group buyer in
+ * its Protected Audience component auction config.
+ */
+export const BUYER_HOSTS_TO_INTEGRATE_BY_SELLER_HOST = new Map([
+  // ad-server -> dsp + dsp-a + dsp-b
+  [AD_SERVER_HOST!, [DSP_HOST!, DSP_A_HOST!, DSP_B_HOST!]],
+  // ssp -> dsp-a + dsp-b
+  [SSP_HOST!, [DSP_A_HOST!, DSP_B_HOST!]],
+  // ssp-a -> dsp + dsp-a
+  [SSP_A_HOST!, [DSP_HOST!, DSP_A_HOST!]],
+  // ssp-b -> dsp + dsp-b
+  [SSP_B_HOST!, [DSP_HOST!, DSP_B_HOST!]],
+]);
