+++ services/ad-tech/src/main.ts
@@ -15,10 +15,19 @@ import express, {Application} from 'express';
 import cors from 'cors';
 
 import {HOSTNAME, PORT} from './lib/constants.js';
-import {WellKnownRouter} from './routes/well-known-router.js';
-import {BuyerRouter} from './routes/buyer-router.js';
-import {CommonRouter} from './routes/common-router.js';
-import {SellerRouter} from './routes/seller-router.js';
+import {BiddingSignalsRouter} from './routes/dsp/bidding-signals-router.js';
+import {BuyerContextualBidderRouter} from './routes/dsp/buyer-contextual-bidder-router.js';
+import {BuyerRouter} from './routes/dsp/buyer-router.js';
+import {CommonRouter} from './routes/common/common-router.js';
+import {ScoringSignalsRouter} from './routes/ssp/scoring-signals-router.js';
+import {SellerContextualBidderRouter} from './routes/ssp/seller-contextual-bidder-router.js';
+import {SellerRouter} from './routes/ssp/seller-router.js';
+import {WellKnownAttributionReportingRouter} from './routes/well-known/well-known-attribution-reporting-router.js';
+import {WellKnownPrivateAggregationRouter} from './routes/well-known/well-known-private-aggregation-router.js';
+import {AdsRouter} from './routes/common/ads-router.js';
+import {ReportRouter} from './routes/common/report-router.js';
+import {AttributionReportingRouter} from './routes/common/attribution-reporting-router.js';
+import {TopicsRouter} from './routes/common/topics-router.js';
 
 const app: Application = express();
 app.use(express.urlencoded({extended: true}));
@@ -28,9 +37,24 @@ app.set('view engine', 'ejs');
 app.set('views', 'src/views');
 
 app.use('/', CommonRouter);
-app.use('/.well-known', WellKnownRouter);
+app.use('/ads', AdsRouter);
+app.use('/reporting', ReportRouter);
+app.use('/attribution', AttributionReportingRouter);
+app.use('/topics', TopicsRouter);
+
+app.use(
+  '/.well-known/attribution-reporting',
+  WellKnownAttributionReportingRouter,
+);
+app.use('/.well-known/private-aggregation', WellKnownPrivateAggregationRouter);
+
 app.use('/dsp', BuyerRouter);
+app.use('/dsp/contextual-bid', BuyerContextualBidderRouter);
+app.use('/dsp/realtime-signals', BiddingSignalsRouter);
+
 app.use('/ssp', SellerRouter);
+app.use('/ssp/contextual-bid', SellerContextualBidderRouter);
+app.use('/ssp/realtime-signals', ScoringSignalsRouter);
 
 app.listen(PORT, function () {
   console.log(HOSTNAME, ' listening on port ', PORT);
