+++ services/ad-tech/src/public/js/display-ad-frame.js
@@ -13,8 +13,8 @@
 
 /**
  * Where is this script used:
- *   This script is included on the ad frame hosted by this ad-tech, and is
- *   executed when an ad is delivered by this ad-tech on a publisher page.
+ *   This script is included in Protected Audience display ads, and is executed
+ *   when an ad is delivered by this ad-tech on a publisher page.
  *
  * What does this script do:
  *   This script triggers the ad beacon that the ad-tech expects to have
@@ -22,14 +22,6 @@
  *   registerAdBeacon().
  */
 (() => {
-  /** Name of the contextual advertiser. */
-  const ADVERTISER_CONTEXTUAL = 'ContextNext';
-  const HOST_CODENAME = ((scriptSrc) => {
-    const currentHost = new URL(scriptSrc).hostname;
-    return currentHost.substring('privacy-sandbox-demos-'.length);
-  })(document.currentScript.src);
-  const CURRENT_URL = new URL(location.href);
-
   /** Triggers all the ad beacons registered in Protected Audience auction. */
   const triggerAdBeacons = () => {
     if (!window.fence?.reportEvent) {
@@ -52,19 +44,18 @@
     });
   };
 
-  /** Adds a description for demonstration purposes. */
+  /** Adds a description for demonstrative purposes. */
   const addDescriptionToAdContainer = () => {
+    const host = new URL(document.currentScript.src).hostname;
     document.addEventListener('DOMContentLoaded', async (e) => {
-      const advertiser = CURRENT_URL.searchParams.get('advertiser');
       const $adLabel = document.getElementById('ad-label');
-      if (ADVERTISER_CONTEXTUAL === advertiser) {
-        $adLabel.innerText = `Contextual ad from ${HOST_CODENAME}`;
-      } else {
-        $adLabel.innerText = `PAAPI ad from ${HOST_CODENAME}`;
+      if ($adLabel) {
+        $adLabel.innerText = `PAAPI ad from ${host}`;
       }
     });
   };
 
+  /** Main function */
   (() => {
     addDescriptionToAdContainer();
     triggerAdBeacons();
