+++ services/ad-tech/src/routes/well-known/well-known-attribution-reporting-router.ts
@@ -13,17 +13,25 @@
 
 import cbor from 'cbor';
 import express, {Request, Response} from 'express';
-import {ReportStore, ReportCategory} from '../controllers/report-store.js';
-import {decodeBucket} from '../lib/arapi.js';
+import {ReportStore, ReportCategory} from '../../controllers/report-store.js';
+import {decodeBucket} from '../../lib/arapi.js';
 
-export const WellKnownRouter = express.Router();
+/**
+ * This router is responsible for handling the well-known endpoints for the
+ * Attribution Reporting API. There are two types reports with this API --
+ * (1) event-level reports, and (2) summary reports or aggregate reports. And
+ * for each report type, we may receive regular reports and debug reports.
+ *
+ * Path: /.well-known/attribution-reporting/
+ */
+export const WellKnownAttributionReportingRouter = express.Router();
 
 // ************************************************************************
-// Attribution Reporting
+// Attribution Reporting: Event-level Reports
 // ************************************************************************
 // Event-level Report: Regular Report
-WellKnownRouter.post(
-  '/attribution-reporting/report-event-attribution',
+WellKnownAttributionReportingRouter.post(
+  '/report-event-attribution',
   async (req: Request, res: Response) => {
     console.log(
       '[ARA] Received event-level report on live endpoint: ',
@@ -39,8 +47,8 @@ WellKnownRouter.post(
 );
 
 // Event-level Report: Debug Report
-WellKnownRouter.post(
-  '/attribution-reporting/debug/report-event-attribution',
+WellKnownAttributionReportingRouter.post(
+  '/debug/report-event-attribution',
   async (req: Request, res: Response) => {
     console.log(
       '[ARA] Received event-level report on debug endpoint: ',
@@ -55,9 +63,12 @@ WellKnownRouter.post(
   },
 );
 
+// ************************************************************************
+// Attribution Reporting: Aggregate Reports
+// ************************************************************************
 // Aggregate Report: Regular Report
-WellKnownRouter.post(
-  '/attribution-reporting/report-aggregate-attribution',
+WellKnownAttributionReportingRouter.post(
+  '/report-aggregate-attribution',
   async (req: Request, res: Response) => {
     const report = req.body;
     report.shared_info = JSON.parse(report.shared_info);
@@ -75,8 +86,8 @@ WellKnownRouter.post(
 );
 
 // Aggregate Report: Debug Report
-WellKnownRouter.post(
-  '/attribution-reporting/debug/report-aggregate-attribution',
+WellKnownAttributionReportingRouter.post(
+  '/debug/report-aggregate-attribution',
   async (req: Request, res: Response) => {
     const debugReport = req.body;
     debugReport.shared_info = JSON.parse(debugReport.shared_info);
@@ -108,76 +119,6 @@ WellKnownRouter.post(
       category: ReportCategory.ARA_AGGREGATE_DEBUG,
       timestamp: Date.now().toString(),
       data: debugReport,
-    });
-    res.sendStatus(200);
-  },
-);
-
-// ************************************************************************
-// Private Aggregation with Shared Storage
-// ************************************************************************
-WellKnownRouter.post(
-  '/private-aggregation/report-shared-storage',
-  async (req: Request, res: Response) => {
-    console.log(
-      '[pAgg+SS] Received aggregatable report on live endpoint: ',
-      req.body,
-    );
-    ReportStore.addReport({
-      category: ReportCategory.PAGG_VIA_SS,
-      timestamp: Date.now().toString(),
-      data: req.body,
-    });
-    res.sendStatus(200);
-  },
-);
-
-WellKnownRouter.post(
-  '/private-aggregation/debug/report-shared-storage',
-  async (req: Request, res: Response) => {
-    console.log(
-      '[pAgg+SS] Received aggregatable report on debug endpoint: ',
-      req.body,
-    );
-    ReportStore.addReport({
-      category: ReportCategory.PAGG_VIA_SS_DEBUG,
-      timestamp: Date.now().toString(),
-      data: req.body,
-    });
-    res.sendStatus(200);
-  },
-);
-
-// ************************************************************************
-// Private Aggregation with Protected Audience
-// ************************************************************************
-WellKnownRouter.post(
-  '/private-aggregation/report-protected-audience',
-  async (req: Request, res: Response) => {
-    console.log(
-      '[pAgg+SS] Received aggregatable report on live endpoint: ',
-      req.body,
-    );
-    ReportStore.addReport({
-      category: ReportCategory.PAGG_VIA_PAAPI,
-      timestamp: Date.now().toString(),
-      data: req.body,
-    });
-    res.sendStatus(200);
-  },
-);
-
-WellKnownRouter.post(
-  '/private-aggregation/debug/report-protected-audience',
-  async (req: Request, res: Response) => {
-    console.log(
-      '[pAgg+SS] Received aggregatable report on debug endpoint: ',
-      req.body,
-    );
-    ReportStore.addReport({
-      category: ReportCategory.PAGG_VIA_PAAPI_DEBUG,
-      timestamp: Date.now().toString(),
-      data: req.body,
     });
     res.sendStatus(200);
   },
