+++ services/ad-tech/src/views/display-ad-frame.ejs
@@ -1,0 +1,61 @@
+<%/* This is used in fenced-frames to deliver ads. */%>
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport"
+        content="width=device-width, initial-scale=1.0" />
+  <title><%= TITLE %></title>
+  <link rel="icon"
+        href="" />
+  <style>
+    body {
+      width: 100vw;
+      height: 100vh;
+      margin: 0;
+      padding: 0;
+      overflow: hidden;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+    }
+
+    #ad-label {
+      position: absolute;
+      top: 0;
+      left: 0;
+      margin: 4px;
+      padding: 4px;
+      background: #3f3f3f;
+      color: white;
+      font-family: monospace;
+    }
+
+    .ad {
+      border: none;
+      margin: 0;
+      padding: 0;
+      height: 100vh;
+      background: white;
+    }
+  </style>
+  <script src="/js/display-ad-frame.js"></script>
+</head>
+
+<body>
+  <p id="ad-label"></p>
+  <a width="300"
+     height="250"
+     target="_blank"
+     rel="noopener noreferrer"
+     href="<%= DESTINATION %>">
+    <!-- Smaller to avoid scrollbar -->
+    <img width="294"
+         height="245"
+         loading="lazy"
+         src="<%= CREATIVE %>">
+  </a>
+</body>
+
+</html>
