+++ services/ad-tech/src/views/dsp/index.ejs
@@ -6,7 +6,7 @@
   <meta charset="UTF-8" />
   <meta name="viewport"
         content="width=device-width, initial-scale=1.0" />
-  <title><%= title %></title>
+  <title><%= TITLE %></title>
   <link rel="stylesheet"
         href="/css/index.css" />
   <link rel="icon"
@@ -14,7 +14,7 @@
 </head>
 
 <body>
-  <h1><%= title %></h1>
+  <h1><%= TITLE %></h1>
   <section>
     <h2>Demand Side Platform</h2>
     <p>DSP provides 3rd party tag which join user into interest group via Protected Audience API.</p>
