+++ services/ad-tech/src/views/ssp/index.ejs
@@ -14,7 +14,7 @@
 </head>
 
 <body>
-  <h1><%= title %></h1>
+  <h1><%= TITLE %></h1>
   <section>
     <h2>Supply Side Platform</h2>
     <p>
