+++ services/dsp-a/README.md
@@ -13,7 +13,7 @@ Serving requested Ads. It includes ads creative (Image, Video etc). Clicking ads
         height="250"
         target="_blank"
         rel="noopener noreferrer"
-        attributionsrc="https://privacy-sandbox-demos-ssp.dev/register-source?advertiser=privacy-sandbox-demos-shop.dev&amp;id=1fa74"
+        attributionsrc="https://privacy-sandbox-demos-ssp.dev/attribution/register-source?advertiser=privacy-sandbox-demos-shop.dev&amp;id=1fa74"
         href="https://privacy-sandbox-demos-shop.dev/items/1fa74"
       >
         <!-- smaller for avoid scrollbar -->
@@ -21,7 +21,7 @@ Serving requested Ads. It includes ads creative (Image, Video etc). Clicking ads
           width="294"
           height="245"
           loading="lazy"
-          attributionsrc="https://privacy-sandbox-demos-ssp.dev/register-source?advertiser=privacy-sandbox-demos-shop.dev&amp;id=1fa74"
+          attributionsrc="https://privacy-sandbox-demos-ssp.dev/attribution/register-source?advertiser=privacy-sandbox-demos-shop.dev&amp;id=1fa74"
           src="
          https://privacy-sandbox-demos-shop.dev/ads/1fa74"
         />
