+++ services/news/src/views/multi-seller.ejs
@@ -21,7 +21,8 @@
       // is included on the page directly. This configuration field should
       // list additional sellers to include in the ad auctions.
       otherSellers: [
-        '<%= SSP_A_ORIGIN %>', 
+        '<%= SSP_ORIGIN %>',
+        '<%= SSP_A_ORIGIN %>',
         '<%= SSP_B_ORIGIN %>',
       ],
       // Ad units to request bids for.
